+++ src/common/js/index.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:collection-timestampable/src/common/js/index.js
+ */
+
+import './timestampable.js';